Restructure: Add InputUsed and reduce use of for
This uses let else or if let instead of the for loops from before, which looks a lot cleaner in this case.

I have also added InputUsed which is a second validation that is required in this case because otherwise the outline color would sometimes change to active color even though it should've stayed on error. This is because on_used got fired before InputError was inserted.